Move interfaces and helpers to separate files
